More warning fixes.

+++ cocos/2d/renderer/CCRenderCommand.cpp
@@ -37,11 +37,11 @@ RenderCommand::~RenderCommand()
 {
 }
 
-void printBits(size_t const size, void const * const ptr)
+void printBits(ssize_t const size, void const * const ptr)
 {
     unsigned char *b = (unsigned char*) ptr;
     unsigned char byte;
-    int i, j;
+    ssize_t i, j;
 
     for (i=size-1;i>=0;i--)
     {
